requirements.txt
langchain==0.1.4
streamlit==1.29.0
openai==1.10.0
tiktoken==0.5.2
python-dotenv==1.0.1
pinecone-client==3.0.1
pypdf==4.0.1
joblib==1.3.2
pandas==2.2.0
scikit-learn==1.4.0
sentence-transformers==2.3.0
langchain-openai==0.0.5